refactor: Reorganize imports and improve documentation formatting
