+++ onvif/services/authenticationbehavior.py
@@ -1,7 +1,8 @@
 """AuthenticationBehavior service implementation."""
 
 from onvif.operator import ONVIFOperator
-from onvif.utils import ONVIFWSDL, ONVIFService
+from onvif.utils.service import ONVIFService
+from onvif.utils.wsdl import ONVIFWSDL
 
 
 # pylint: disable=invalid-name
